Refactor component structure by moving common components to a centralized directory. Updated import paths in layout and loading files to reflect the new structure. Added new common components including Header, Footer, LoadingScreen, and GroupSummary for improved organization and maintainability.
