+++ src/app/_components/features/expense/ExpenseForm/ExpenseDetails.tsx
@@ -8,7 +8,7 @@ import {
   SelectValue,
 } from "~/components/ui/select";
 import { IndianRupee, Receipt, Users } from "lucide-react";
-import { type Person, getPersonInitials } from "~/app/_components/group/utils";
+import { type Person, getPersonInitials } from "~/app/_components/features/group-management/utils";
 import { PeopleSelection } from "./PeopleSelection";
 
 interface ExpenseDetailsProps {
