+++ src/app/invite/[token]/loading.tsx
@@ -1,4 +1,4 @@
-import { LoadingScreen } from "../../_components/LoadingScreen";
+import { LoadingScreen } from "../../_components/common/LoadingScreen";
 
 export default function Loading() {
   return <LoadingScreen />;
