+++ src/app/invite/[token]/page.tsx
@@ -1,4 +1,4 @@
-import { InvitePageClient } from "~/app/_components/invite/InvitePageClient";
+import { InvitePageClient } from "~/app/_components/features/auth/invite/InvitePageClient";
 
 export default async function InvitePage({
   params,
